+++ lib/forge_sdk/protocol/display/indexed_transaction.ex
@@ -4,11 +4,11 @@ defimpl ForgeSdk.Display, for: ForgeAbi.IndexedTransaction do
   """
   alias ForgeSdk.Display
 
-  def display(data, _expand? \\ false) do
+  def display(data, expand? \\ false) do
     basic = Map.from_struct(data)
 
     Map.merge(basic, %{
-      tx: Display.display(basic.tx),
+      tx: Display.display(basic.tx, expand?),
       code: ForgeAbi.StatusCode.value(basic.code)
     })
   end